+++ multisig/proposals/zksyncsepolia/vip-016/index.ts
@@ -4,7 +4,7 @@ export const ZKSYNCSEPOLIA_ACM = "0xD07f543d47c3a8997D6079958308e981AC14CD01";
 export const ZKSYNCSEPOLIA_NORMAL_TIMELOCK = "0x1730527a0f0930269313D77A317361b42971a67E";
 const DEFAULT_ADMIN_ROLE = "0x0000000000000000000000000000000000000000000000000000000000000000";
 
-export const vip010 = () => {
+export const vip016 = () => {
   return makeProposal([
     {
       target: ZKSYNCSEPOLIA_ACM,
@@ -13,4 +13,4 @@ export const vip010 = () => {
     },
   ]);
 };
-export default vip010;
+export default vip016;
