--- simulations/vip-395/abi/OmnichainExecutorOwner_ABI.json
@@ -1,145 +1,0 @@
-[
-  {
-    "inputs": [{ "internalType": "address", "name": "omnichainGovernanceExecutor_", "type": "address" }],
-    "stateMutability": "nonpayable",
-    "type": "constructor"
-  },
-  {
-    "inputs": [
-      { "internalType": "address", "name": "sender", "type": "address" },
-      { "internalType": "address", "name": "calledContract", "type": "address" },
-      { "internalType": "string", "name": "methodSignature", "type": "string" }
-    ],
-    "name": "Unauthorized",
-    "type": "error"
-  },
-  { "inputs": [], "name": "ZeroAddressNotAllowed", "type": "error" },
-  {
-    "anonymous": false,
-    "inputs": [
-      { "indexed": true, "internalType": "string", "name": "signature", "type": "string" },
-      { "indexed": false, "internalType": "bool", "name": "active", "type": "bool" }
-    ],
-    "name": "FunctionRegistryChanged",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [{ "indexed": false, "internalType": "uint8", "name": "version", "type": "uint8" }],
-    "name": "Initialized",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      { "indexed": false, "internalType": "address", "name": "oldAccessControlManager", "type": "address" },
-      { "indexed": false, "internalType": "address", "name": "newAccessControlManager", "type": "address" }
-    ],
-    "name": "NewAccessControlManager",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      { "indexed": true, "internalType": "address", "name": "previousOwner", "type": "address" },
-      { "indexed": true, "internalType": "address", "name": "newOwner", "type": "address" }
-    ],
-    "name": "OwnershipTransferStarted",
-    "type": "event"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      { "indexed": true, "internalType": "address", "name": "previousOwner", "type": "address" },
-      { "indexed": true, "internalType": "address", "name": "newOwner", "type": "address" }
-    ],
-    "name": "OwnershipTransferred",
-    "type": "event"
-  },
-  { "stateMutability": "nonpayable", "type": "fallback" },
-  {
-    "inputs": [],
-    "name": "OMNICHAIN_GOVERNANCE_EXECUTOR",
-    "outputs": [{ "internalType": "contract IOmnichainGovernanceExecutor", "name": "", "type": "address" }],
-    "stateMutability": "view",
-    "type": "function"
-  },
-  { "inputs": [], "name": "acceptOwnership", "outputs": [], "stateMutability": "nonpayable", "type": "function" },
-  {
-    "inputs": [],
-    "name": "accessControlManager",
-    "outputs": [{ "internalType": "contract IAccessControlManagerV8", "name": "", "type": "address" }],
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "inputs": [{ "internalType": "bytes4", "name": "", "type": "bytes4" }],
-    "name": "functionRegistry",
-    "outputs": [{ "internalType": "string", "name": "", "type": "string" }],
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "inputs": [{ "internalType": "address", "name": "accessControlManager_", "type": "address" }],
-    "name": "initialize",
-    "outputs": [],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [],
-    "name": "owner",
-    "outputs": [{ "internalType": "address", "name": "", "type": "address" }],
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "inputs": [],
-    "name": "pendingOwner",
-    "outputs": [{ "internalType": "address", "name": "", "type": "address" }],
-    "stateMutability": "view",
-    "type": "function"
-  },
-  { "inputs": [], "name": "renounceOwnership", "outputs": [], "stateMutability": "nonpayable", "type": "function" },
-  {
-    "inputs": [{ "internalType": "address", "name": "accessControlManager_", "type": "address" }],
-    "name": "setAccessControlManager",
-    "outputs": [],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      { "internalType": "uint16", "name": "srcChainId_", "type": "uint16" },
-      { "internalType": "bytes", "name": "srcAddress_", "type": "bytes" }
-    ],
-    "name": "setTrustedRemoteAddress",
-    "outputs": [],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [{ "internalType": "address", "name": "newOwner_", "type": "address" }],
-    "name": "transferBridgeOwnership",
-    "outputs": [],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [{ "internalType": "address", "name": "newOwner", "type": "address" }],
-    "name": "transferOwnership",
-    "outputs": [],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      { "internalType": "string[]", "name": "signatures_", "type": "string[]" },
-      { "internalType": "bool[]", "name": "active_", "type": "bool[]" }
-    ],
-    "name": "upsertSignature",
-    "outputs": [],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  }
-]
